codegen: Go 1.23 alias support
